+++ .gitignore
@@ -1,5 +1,6 @@
 .vscode
 build/
 **/gorestldap
+!cmd/**
 **/*.so
 **/gorestldap.tar.gz
