+++ internal/http/http.go
@@ -1,4 +1,4 @@
-package main
+package http
 
 import (
 	"strings"
@@ -8,7 +8,18 @@ import (
 	"github.com/valyala/fasthttp"
 )
 
-func handleCallback(ctx *fasthttp.RequestCtx, cfg *Config, ticker *time.Ticker) {
+func NewServer(callbackAuthToken string, updateInterval time.Duration, ticker *time.Ticker) *fasthttp.Server {
+	return &fasthttp.Server{
+		ReadTimeout:  5 * time.Second,
+		WriteTimeout: 15 * time.Second,
+		IdleTimeout:  120 * time.Second,
+		Handler: func(ctx *fasthttp.RequestCtx) {
+			handleCallback(ctx, callbackAuthToken, updateInterval, ticker)
+		},
+	}
+}
+
+func handleCallback(ctx *fasthttp.RequestCtx, callbackAuthToken string, updateInterval time.Duration, ticker *time.Ticker) {
 	switch path := string(ctx.Path()); path {
 	case "/callback":
 		log.Debug("new callback request")
@@ -20,14 +31,14 @@ func handleCallback(ctx *fasthttp.RequestCtx, cfg *Config, ticker *time.Ticker) 
 
 		hAuthBytes := ctx.Request.Header.Peek("Authorization")
 		cbToken := strings.TrimPrefix(string(hAuthBytes), "Token ")
-		if cbToken != cfg.CallbackAuthToken {
+		if cbToken != callbackAuthToken {
 			ctx.SetStatusCode(fasthttp.StatusUnauthorized)
 			return
 		}
 
 		ticker.Reset(time.Millisecond)
 		<-ticker.C
-		ticker.Reset(cfg.UpdateInterval)
+		ticker.Reset(updateInterval)
 	default:
 		ctx.Redirect("/callback", fasthttp.StatusMovedPermanently)
 	}
