+++ internal/ldap/ldap_helpers.go
@@ -1,4 +1,4 @@
-package main
+package ldap
 
 import (
 	"errors"
